Convert CommonJS to ESModule
