+++ backend/package.json
@@ -2,6 +2,7 @@
   "name": "reserve.utsa",
   "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "start": "node ./bin/www"
   },
