+++ backend/routes/reserve/get_items.js
@@ -1,5 +1,6 @@
-const express = require('express')
-const { items } = require('../../models/mongodb')
+import express from 'express'
+import { items } from '../../models/mongodb.js'
+
 const router = express.Router()
 
 /**
@@ -26,13 +27,13 @@ const router = express.Router()
  *                     $ref: '#/components/schemas/Item'
  */
 router.get('/items', async function(req, res, next) {
-    const data = await items
-        .find({})
-        .project({ _id: 1, name: 1, quantity: 1, exception_time: 1 })
-        .toArray( function(err, results) {
-            console.log(results)
-        })
-    res.json({ data: data })
+  const data = await items
+    .find({})
+    .project({ _id: 1, name: 1, quantity: 1, exception_time: 1 })
+    .toArray( function(err, results) {
+      console.log(results)
+    })
+  res.json({ data: data })
 })
 
-module.exports = router
+export default router
