+++ backend/routes/reserve/index.js
@@ -1,17 +1,50 @@
-const reserve_router = require('express').Router()
+import express from 'express'
+const reserve_router = express.Router()
+// const reserve_router = require('express').Router()
 
 const root = '/reserve'
+function useRouter(router) {
+    reserve_router.use(root, router)
+}
 
-reserve_router.use(root, require('./GET_items'))
-reserve_router.use(root, require('./GET_spaces'))
-reserve_router.use(root, require('./GET_integral_space_availability'))
-reserve_router.use(root, require('./GET_interval_space_availability'))
-reserve_router.use(root, require('./GET_integral_item_availability'))
-reserve_router.use(root, require('./GET_interval_item_availability'))
-reserve_router.use(root, require('./GET_reservation'))
-reserve_router.use(root, require('./POST_reservation'))
-reserve_router.use(root, require('./PUT_reservation'))
-reserve_router.use(root, require('./DELETE_reservation'))
-reserve_router.use(root, require('./GET_verify'))
+async function useRouters() {
+    reserve_router.use(root, (await import('./get_items.js')).default)
+    reserve_router.use(root, (await import('./get_spaces.js')).default)
+    reserve_router.use(root, (await import('./get_integral_space_availability.js')).default)
+    reserve_router.use(root, (await import('./get_interval_space_availability.js')).default)
+    reserve_router.use(root, (await import('./get_integral_item_availability.js')).default)
+    reserve_router.use(root, (await import('./get_interval_item_availability.js')).default)
+    reserve_router.use(root, (await import('./get_reservation.js')).default)
+    reserve_router.use(root, (await import('./post_reservation.js')).default)
+    reserve_router.use(root, (await import('./put_reservation.js')).default)
+    reserve_router.use(root, (await import('./delete_reservation.js')).default)
+    reserve_router.use(root, (await import('./get_verify.js')).default)
+}
 
-module.exports = reserve_router;
+// import('./get_items.js').then(useRouter)
+// import('./get_spaces.js').then(useRouter)
+// import('./get_integral_space_availability.js').then(useRouter)
+// import('./get_interval_space_availability.js').then(useRouter)
+// import('./get_integral_item_availability.js').then(useRouter)
+// import('./get_interval_item_availability.js').then(useRouter)
+// import('./get_reservation.js').then(useRouter)
+// import('./post_reservation.js').then(useRouter)
+// import('./put_reservation.js').then(useRouter)
+// import('./delete_reservation.js').then(useRouter)
+// import('./get_verify.js').then(useRouter)
+
+// reserve_router.use(root, require('./get_items'))
+// reserve_router.use(root, require('./get_spaces'))
+// reserve_router.use(root, require('./get_integral_space_availability'))
+// reserve_router.use(root, require('./get_interval_space_availability'))
+// reserve_router.use(root, require('./get_integral_item_availability'))
+// reserve_router.use(root, require('./get_interval_item_availability'))
+// reserve_router.use(root, require('./get_reservation'))
+// reserve_router.use(root, require('./post_reservation'))
+// reserve_router.use(root, require('./put_reservation'))
+// reserve_router.use(root, require('./delete_reservation'))
+// reserve_router.use(root, require('./get_verify'))
+
+useRouters()
+
+export default reserve_router
